+++ web_poet/mixins.py
@@ -9,7 +9,6 @@ class ResponseShortcutsMixin:
 
     It requires "response" attribute to be present.
     """
-    _cached_selector = None
     _cached_base_url = None
 
     @property
@@ -20,15 +19,14 @@ class ResponseShortcutsMixin:
     @property
     def html(self):
         """Shortcut to HTML Response's content."""
-        return self.response.html
+        return self.response.text
 
     @property
     def selector(self) -> parsel.Selector:
         """``parsel.Selector`` instance for the HTML Response."""
-        if self._cached_selector is None:
-            self._cached_selector = parsel.Selector(self.html)
-
-        return self._cached_selector
+        # TODO: when dropping Python 3.7 support,
+        #  implement it using typing.Protocol
+        return self.response.selector  # type: ignore
 
     def xpath(self, query, **kwargs):
         """Run an XPath query on a response, using :class:`parsel.Selector`."""
@@ -41,6 +39,7 @@ class ResponseShortcutsMixin:
     @property
     def base_url(self) -> str:
         """Return the base url of the given response"""
+        # FIXME: move it to HttpResponse
         if self._cached_base_url is None:
             text = self.html[:4096]
             self._cached_base_url = get_base_url(text, self.url)
@@ -49,5 +48,5 @@ class ResponseShortcutsMixin:
     def urljoin(self, url: str) -> str:
         """Convert url to absolute, taking in account
         url and baseurl of the response"""
+        # FIXME: move it to HttpResponse
         return urljoin(self.base_url, url)
-